Connection: Move update_user_token to the Connection package (#14001)
* Connection: Move update_user_token to the Connection package

Move the update_user_token method to the Connection page and deprecate
the existing update_user_token method in the Jetpack class. This change
helps to separate the Connection package from the plugin.

* Connection: Move invocations of update_user_token to the Utils class

The update_user_token method has been moved to the Connection package's
Utils class, so move all invocations of this method to the Utils class.