Connection\Utils: Add a new function which provides the Jetpack API version (#14337)
* Connection: Add a new function which provides the Jetpack API version

Add a new function which provides the Jetpack API version to the Util class. It
will return the JETPACK__API_VERSION constant if it's available. Otherwise, it will
return a default value of 1.

Use this new function to obtain the Jetpack API version in the Manager and Client
classes.

* Connection Utils: add unit tests for Utils::get_jetpack_api_version()

Add two unit tests for the new Utils::get_jetpack_api_version() method.