Fix remaining phpcs warnings in most of requirelist (#17694)
Skipped packages/analyzer, there's too much going on there.

The most interesting bit is the switching of abtest to use WorDBless so
it can pick up the WP_Error definition it needs instead of trying to
locally delcare it in its tests which results in a phpcs complaint about
the redefined class.

Also account for two files in the requirelist that had been moved.

I'm tempted to swtich the requirelist to an "excludelist" so we can more
easily exclude things like packages/analyzer and don't have to worry
about tracking new files and renames of files that are supposed to be
clean, but I'll leave that for a future PR.

Committed via a GitHub action: https://github.com/automattic/jetpack/runs/366746452